+++ napari/plugins/_tests/test_builtin_get_writer.py
@@ -1,30 +1,14 @@
 import importlib
 import os
-import numpy as np
 import pytest
 from napari.plugins.exceptions import PluginCallError
 from napari.plugins._builtins import napari_get_writer
-from napari.layers import Image, Points
+from napari.plugins._tests.fixtures.layer_data import (  # noqa: F401
+    layer_data_and_types,
+)
 
 
-@pytest.fixture
-def layer_data_and_types():
-    layers = [
-        Image(np.random.rand(20, 20)),
-        Image(np.random.rand(20, 20)),
-        Points(np.random.rand(20, 2)),
-        Points(
-            np.random.rand(20, 2), properties={'values': np.random.rand(20)}
-        ),
-    ]
-    extensions = ['.tif', '.tif', '.csv', '.csv']
-    layer_data = [l.as_layer_data_tuple() for l in layers]
-    layer_types = [ld[2] for ld in layer_data]
-    filenames = [l.name + e for l, e in zip(layers, extensions)]
-    return layer_data, layer_types, filenames
-
-
-def test_get_writer(tmpdir, layer_data_and_types):
+def test_get_writer(tmpdir, layer_data_and_types):  # noqa: F811
     """Test writing layers data."""
     # make individual write layer builtin plugins get called first
     from napari.plugins import plugin_manager
@@ -32,7 +16,7 @@ def test_get_writer(tmpdir, layer_data_and_types):
     plugin_manager.hooks.napari_write_image.bring_to_front(['builtins'])
     plugin_manager.hooks.napari_write_points.bring_to_front(['builtins'])
 
-    layer_data, layer_types, filenames = layer_data_and_types
+    _, layer_data, layer_types, filenames = layer_data_and_types
 
     path = os.path.join(tmpdir, 'layers_folder')
 
@@ -58,7 +42,7 @@ def test_get_writer(tmpdir, layer_data_and_types):
     assert set(os.listdir(tmpdir)) == set(['layers_folder'])
 
 
-def test_get_writer_bad_plugin(tmpdir, layer_data_and_types):
+def test_get_writer_bad_plugin(tmpdir, layer_data_and_types):  # noqa: F811
     """Test writing layers data."""
     # make individual write layer builtin plugins get called first
     from napari.plugins import plugin_manager
@@ -69,7 +53,7 @@ def test_get_writer_bad_plugin(tmpdir, layer_data_and_types):
     plugin_manager.register(bad)
     plugin_manager.hooks.napari_write_points.bring_to_front([bad_plugin_path])
 
-    layer_data, layer_types, filenames = layer_data_and_types
+    _, layer_data, layer_types, filenames = layer_data_and_types
 
     path = os.path.join(tmpdir, 'layers_folder')
